yashasproject/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "jetson-autopilot"
version = "2.0.0"
description = "Self-driving toy car system using CNN for NVIDIA Jetson"
readme = "README.md"
requires-python = ">=3.7"
license = {text = "MIT"}
authors = [
    {name = "Greg Surma"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "torch>=1.9.0",
    "torchvision>=0.10.0",
    "numpy>=1.19.0",
    "opencv-python>=4.5.0",
    "Pillow>=8.0.0",
    "tqdm>=4.60.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=6.0.0",
    "pytest-cov>=2.0.0",
    "black>=21.0.0",
    "mypy>=0.900",
    "ruff>=0.0.200",
]
notebook = [
    "jupyter>=1.0.0",
    "ipywidgets>=7.6.0",
    "matplotlib>=3.4.0",
]

[project.scripts]
jetson-autopilot = "jetson.cli:main"

[project.urls]
Homepage = "https://github.com/gsurma/jetson"
Repository = "https://github.com/gsurma/jetson"

[tool.setuptools.packages.find]
include = ["jetson*"]

[tool.black]
line-length = 100
target-version = ["py37", "py38", "py39", "py310", "py311"]

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.7"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
